fix: address PR review round 4 (cursor + codex 4 P2 + 1 CI flake)
Round 4 of bot review. All findings valid; one was a CI flake on a
pre-existing test that my Phase 2 changes destabilized via slower CI
timing on Blacksmith runners.

1. **CI flake (Blacksmith): `RebalanceIncentiveUpdated writes newIncentive
   to Pool.rebalanceReward` timed out at 10s.** The test's `seedFpmmPool`
   helper triggered the FPMM factory handler which made a real RPC
   `decimals` call (test addresses don't exist on-chain → fall-through →
   ERC20 fallback hits Forno → slow on CI). Mocked
   `_setMockERC20Decimals` for both test tokens in `seedFpmmPool` so the
   test is deterministic. Pre-existing latent flake unrelated to Phase 2,
   but it broke on this PR so fixing here.

2. **Cursor pool.ts:736 — `oracleDelta` clobbers healed `referenceRateFeedID`
   in VirtualPoolDeployed.** The post-upsertPool `mirrorFeedIdToPool` was
   rescuing the bug from upsertPool's spread chain, not preventing it.
   Fix: remove `referenceRateFeedID` from `DEFAULT_ORACLE_FIELDS` (it's a
   static field, not an oracle delta). Add a dedicated
   `referenceRateFeedID` param to upsertPool. FPMM factory passes it
   directly (no spread). Self-heal pipeline writes via the same param.
   Now the spread chain doesn't have the power to clobber.

3. **Codex pool-header.tsx:59 — header treats zero-feed as Active.**
   `statusKey()` would return "active" for a stub-row VirtualPool whose
   `referenceRateFeedID` is the zero-address sentinel, while the panel
   correctly rendered the syncing note — operator sees "Active" + syncing
   note simultaneously. Fix: header `statusKey()` now checks zero-feed
   AFTER `isDeprecated` (deprecated wins) and routes to "loading" state.

4. **Codex biPoolManager.ts:253 — skip BiPoolManager RPC backfills during
   preload.** Both multichain configs have `preload_handlers: true`, so
   every event runs twice. The non-cached `poolExchangeEffect` was firing
   on both passes. Added `if (context.isPreload) { ...entity warming...;
   return; }` guards to ExchangeCreated, ExchangeDestroyed, BucketsUpdated,
   SpreadUpdated handlers. Preload now does just the entity reads
   (BiPoolExchange.get + Pool.getWhere) for cache warming; processing
   pass owns all RPC and writes.

5. **Codex pool-state.ts:1091 — split BiPoolManager fetchers to stay
   under 1000-line cap.** Repo `AGENTS.md` enforces a hard 1000-line cap
   on source files; my Phase 2 additions pushed `pool-state.ts` to 1107.
   Moved BiPoolManager / VirtualPool fetchers (`fetchPoolExchange`,
   `fetchVirtualPoolExchangeId`, `extractVpExchangeIdFromBytecode`,
   their mocks + types) into a new `src/rpc/biPoolManager.ts` (~200
   lines). `pool-state.ts` is back to 909 lines. Re-exports in
   `src/rpc.ts` updated; effects.ts imports from the new module;
   test imports updated.

605 indexer + 1967 dashboard tests pass. Both mainnet + testnet codegens
clean. Both files under the 1000-line cap.
